Move Wikipedia search to separate file
