+++ packages/aiqtoolkit_langchain/src/aiq/plugins/langchain/tools/register.py
@@ -21,3 +21,4 @@
 
 from . import tavily_internet_search
 from . import code_generation_tool
+from . import wikipedia_search
